Rolling back hierarchy change
