+++ tests/collections/tts/test_tts_exportables.py
@@ -54,8 +54,7 @@ def radtts_model():
     model = RadTTSModel(cfg=cfg.model)
     app_state.is_model_being_restored = False
     model.eval()
-    model.export_config['enable_ragged_batches'] = True
-    model.export_config['enable_volume'] = True
+    model.set_export_config({'enable_ragged_batches': 'True', 'enable_volume': 'True'})
     return model
